remove multierror_scripts, rename _core/_multierror.py to _core/_concat_tb.py
