--- src/trio/_core/_tests/test_multierror_scripts/__init__.py
@@ -1,3 +1,0 @@
-# This isn't really a package, everything in here is a standalone script. This
-# __init__.py is just to fool setuptools' find = {namespaces = false} into
-# actually installing the things.
