--- src/trio/_core/_tests/test_multierror_scripts/_common.py
@@ -1,7 +1,0 @@
-# https://coverage.readthedocs.io/en/latest/subprocess.html
-try:
-    import coverage
-except ImportError:  # pragma: no cover
-    pass
-else:
-    coverage.process_startup()
